backup/FunctionAppVersion/requirements.txt
pyodbc
google-api-python-client
google-auth-httplib2
google-auth-oauthlib